backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
google-generativeai==0.8.3
groq==0.11.0
sentence-transformers==3.3.1
faiss-cpu==1.9.0
numpy==1.26.4
pypdf==5.1.0
python-multipart==0.0.12
